test(body-wire): -Wsign-compare 경고 셋
로컬 MinGW 는 안 내고 CI 의 gcc 만 냈다. 다섯 번째 패턴 — 내가 굽는
것이 CI 가 굽는 것보다 좁다. 이번에는 경고였지만 같은 자리에서
NULL/stddef.h 는 빌드 실패였다.

앞으로 이 저장소의 테스트를 고칠 때는 -Wsign-compare 를 명시해서
한 번 더 굽는다.

+++ test/test_body_wire.c
@@ -69,7 +69,7 @@ static void test_masks_never_overlap(void) {
             if (x->mux_byte != FSD_BODY_WIRE_NO_MUX && x->mux_byte == y->mux_byte &&
                 x->mux_value != y->mux_value)
                 continue;
-            for (int i = 0; i < FSD_BODY_WIRE_MAX_DLC; i++)
+            for (int i = 0; i < (int)FSD_BODY_WIRE_MAX_DLC; i++)
                 CHECK((x->payload[i] & y->payload[i]) == 0,
                       "%s and %s both claim byte %d (0x%02X & 0x%02X)",
                       fsd_body_action_str((FsdBodyAction)a),
@@ -105,7 +105,7 @@ static void test_unknown_frames_have_no_row(void) {
     // its bits are not. It may change nothing.
     const FsdBodyWire *p = fsd_body_wire(FSD_ACT_SEAT_PASSENGER);
     CHECK(p != NULL, "passenger row exists");
-    for (int i = 0; p && i < FSD_BODY_WIRE_MAX_DLC; i++)
+    for (int i = 0; p && i < (int)FSD_BODY_WIRE_MAX_DLC; i++)
         CHECK(p->payload[i] == 0, "passenger payload byte %d is empty (rules 2/11 uncaptured)", i);
 
     uint8_t out[8];
@@ -300,7 +300,7 @@ static void test_masks_are_pinned(void) {
         CHECK(w->can_id == expect[k].id, "%s frame is 0x%03X",
               fsd_body_action_str(expect[k].a), (unsigned)expect[k].id);
         CHECK(w->dlc == expect[k].dlc, "%s length", fsd_body_action_str(expect[k].a));
-        for (int i = 0; i < FSD_BODY_WIRE_MAX_DLC; i++)
+        for (int i = 0; i < (int)FSD_BODY_WIRE_MAX_DLC; i++)
             CHECK(w->payload[i] == expect[k].payload[i],
                   "%s byte %d mask: row 0x%02X, pinned 0x%02X",
                   fsd_body_action_str(expect[k].a), i, w->payload[i], expect[k].payload[i]);
